requirements.txt
Flask==2.3.3
line-bot-sdk>=3.14.5
python-dotenv==1.0.0
pytest==7.4.0
requests>=2.31.0
